genai-rag-web-ingestion/embedding/requirement.txt
python-dotenv
azure-storage-blob
azure-search-documents
openai
